unsigned indices

+++ include/tepp/simple_vector.h
@@ -13,10 +13,10 @@ namespace te
 	struct simple_vector
 	{
 		T* data = nullptr;
-		size_t capacity = 0;
-		size_t size = 0;
+		int64_t capacity = 0;
+		int64_t size = 0;
 
-		inline T& operator[](size_t i) {
+		inline T& operator[](int64_t i) {
 			assert(i >= 0);
 			assert(i < this->size);
 
@@ -59,12 +59,12 @@ namespace te
 	private:
 		inline void ensure_capacity() {
 			if (this->size == this->capacity) {
-				this->capacity = std::max(size_t(1), this->capacity * 2);
+				this->capacity = std::max(int64_t(1), this->capacity * 2);
 				auto newBuffer = new T[this->capacity];
 
 				auto it = this->data;
 				auto jt = newBuffer;
-				for (size_t i = 0; i < this->size; i++, it++, jt++) {
+				for (int64_t i = 0; i < this->size; i++, it++, jt++) {
 					*jt = std::move(*it);
 				}
 				delete[] this->data;
@@ -152,7 +152,7 @@ namespace te
 		simple_vector& operator=(simple_vector const& other) = delete;
 
 		simple_vector() = default;
-		simple_vector(size_t capacity_) {
+		simple_vector(int64_t capacity_) {
 			this->capacity = capacity_;
 			this->data = new T[capacity];
 		}
@@ -170,5 +170,4 @@ namespace te
 			}
 		}
 	};
-
 }
